move device tests into test_spyre test class/file
